+++ LSTMText/requirements.txt
@@ -1,0 +1,3 @@
+numpy
+six
+shutil
