+++ Transformer/imdb/utils.py
@@ -1,0 +1,10 @@
+import sys
+
+sys.path.append("../../")
+from imdb.utils import pad_sequences, load_imdb_data, colored_string
+
+__all__ = [
+    "pad_sequences",
+    "load_imdb_data",
+    "colored_string"
+]
