--- Transformer/models/transformer.py
@@ -1,407 +1,0 @@
-import copy
-from typing import Optional, Any
-
-import oneflow as flow
-
-from oneflow import Tensor
-import oneflow.F as F
-from oneflow.nn import Module, ModuleList, Dropout, Linear, LayerNorm
-from .multihead_attention import MultiheadAttention
-from oneflow.nn.init import xavier_uniform_
-
-
-class Transformer(Module):
-    r"""
-    Args:
-        d_model: the number of expected features in the encoder/decoder inputs (default=512).
-        nhead: the number of heads in the multiheadattention models (default=8).
-        num_encoder_layers: the number of sub-encoder-layers in the encoder (default=6).
-        num_decoder_layers: the number of sub-decoder-layers in the decoder (default=6).
-        dim_feedforward: the dimension of the feedforward network model (default=2048).
-        dropout: the dropout value (default=0.1).
-        activation: the activation function of encoder/decoder intermediate layer, relu or gelu (default=relu).
-        custom_encoder: custom encoder (default=None).
-        custom_decoder: custom decoder (default=None).
-        layer_norm_eps: the eps value in layer normalization components (default=1e-5).
-        batch_first: If ``True``, then the input and output tensors are provided
-            as (batch, seq, feature). Default: ``False`` (seq, batch, feature).
-        norm_first: if ``True``, encoder and decoder layers will perform LayerNorms before
-            other attention and feedforward operations, otherwise after. Default: ``False`` (after).
-    """
-
-    def __init__(self, d_model: int = 512, nhead: int = 8, num_encoder_layers: int = 6,
-                 num_decoder_layers: int = 6, dim_feedforward: int = 2048, dropout: float = 0.1,
-                 activation: str = "relu", custom_encoder: Optional[Any] = None, custom_decoder: Optional[Any] = None,
-                 layer_norm_eps: float = 1e-5, batch_first: bool = False, norm_first: bool = False,
-                 device=None, dtype=None) -> None:
-        # factory_kwargs = {'device': device, 'dtype': dtype}
-        super(Transformer, self).__init__()
-
-        if custom_encoder is not None:
-            self.encoder = custom_encoder
-        else:
-            # encoder_layer = TransformerEncoderLayer(d_model, nhead, dim_feedforward, dropout,
-            #                                        activation, layer_norm_eps, batch_first, norm_first)
-            encoder_norm = LayerNorm(d_model, eps=layer_norm_eps)
-            # self.encoder = TransformerEncoder(
-            #    encoder_layer, num_encoder_layers, encoder_norm)
-            self.encoder = TransformerEncoder(num_encoder_layers, d_model, nhead, dim_feedforward, dropout,
-                                              activation, layer_norm_eps, batch_first, norm_first, encoder_norm)
-
-        if custom_decoder is not None:
-            self.decoder = custom_decoder
-        else:
-            # decoder_layer = TransformerDecoderLayer(d_model, nhead, dim_feedforward, dropout,
-            #                                        activation, layer_norm_eps, batch_first, norm_first)
-            decoder_norm = LayerNorm(d_model, eps=layer_norm_eps)
-            # self.decoder = TransformerDecoder(
-            #    decoder_layer, num_decoder_layers, decoder_norm)
-            self.decoder = TransformerDecoder(num_decoder_layers, d_model, nhead, dim_feedforward, dropout,
-                                              activation, layer_norm_eps, batch_first, norm_first, decoder_norm)
-
-        self._reset_parameters()
-
-        self.d_model = d_model
-        self.nhead = nhead
-
-        self.batch_first = batch_first
-
-    def forward(self, src: Tensor, tgt: Tensor, src_mask: Optional[Tensor] = None, tgt_mask: Optional[Tensor] = None,
-                memory_mask: Optional[Tensor] = None, src_key_padding_mask: Optional[Tensor] = None,
-                tgt_key_padding_mask: Optional[Tensor] = None,
-                memory_key_padding_mask: Optional[Tensor] = None) -> Tensor:
-        r"""Take in and process masked source/target sequences.
-
-        Args:
-            src: the sequence to the encoder (required).
-            tgt: the sequence to the decoder (required).
-            src_mask: the additive mask for the src sequence (optional).
-            tgt_mask: the additive mask for the tgt sequence (optional).
-            memory_mask: the additive mask for the encoder output (optional).
-            src_key_padding_mask: the ByteTensor mask for src keys per batch (optional).
-            tgt_key_padding_mask: the ByteTensor mask for tgt keys per batch (optional).
-            memory_key_padding_mask: the ByteTensor mask for memory keys per batch (optional).
-
-        Shape:
-            - src: :math:`(S, N, E)`, `(N, S, E)` if batch_first.
-            - tgt: :math:`(T, N, E)`, `(N, T, E)` if batch_first.
-            - src_mask: :math:`(S, S)`.
-            - tgt_mask: :math:`(T, T)`.
-            - memory_mask: :math:`(T, S)`.
-            - src_key_padding_mask: :math:`(N, S)`.
-            - tgt_key_padding_mask: :math:`(N, T)`.
-            - memory_key_padding_mask: :math:`(N, S)`.
-
-            - output: :math:`(T, N, E)`, `(N, T, E)` if batch_first.
-
-            where S is the source sequence length, T is the target sequence length, N is the
-            batch size, E is the feature number
-        """
-
-        if not self.batch_first and src.size(1) != tgt.size(1):
-            raise RuntimeError("the batch number of src and tgt must be equal")
-        elif self.batch_first and src.size(0) != tgt.size(0):
-            raise RuntimeError("the batch number of src and tgt must be equal")
-
-        if src.size(2) != self.d_model or tgt.size(2) != self.d_model:
-            raise RuntimeError(
-                "the feature number of src and tgt must be equal to d_model")
-
-        memory = self.encoder(src, src_mask, src_key_padding_mask)
-        output = self.decoder(tgt, memory, tgt_mask, memory_mask,
-                              tgt_key_padding_mask, memory_key_padding_mask)
-        return output
-
-    def generate_square_subsequent_mask(self, sz: int) -> Tensor:
-        r"""Generate a square mask for the sequence. The masked positions are filled with float('-inf').
-            Unmasked positions are filled with float(0.0).
-        """
-        return flow.triu(flow.ones((sz, sz))*float('-inf'), diagonal=1)
-
-    def _reset_parameters(self):
-        r"""Initiate parameters in the transformer model."""
-
-        for p in self.parameters():
-            if p.dim() > 1:
-                xavier_uniform_(p)
-
-
-class TransformerEncoder(Module):
-    r"""TransformerEncoder is a stack of N encoder layers
-
-    Args:
-        encoder_layer: an instance of the TransformerEncoderLayer() class (required).
-        num_layers: the number of sub-encoder-layers in the encoder (required).
-        norm: the layer normalization component (optional).
-    """
-    __constants__ = ['norm']
-
-    #def __init__(self, encoder_layer, num_layers, norm=None):
-    def __init__(self, num_layers, d_model, nhead, dim_feedforward=2048, dropout=0.1, activation="relu",
-                 layer_norm_eps=1e-5, batch_first=False, norm_first=False, norm=None,
-                 device=None, dtype=None):
-        super(TransformerEncoder, self).__init__()
-        #self.layers = _get_clones(encoder_layer, num_layers)
-        self.layers = ModuleList([TransformerEncoderLayer(d_model, nhead, dim_feedforward, dropout,
-                                                          activation, layer_norm_eps, batch_first, norm_first,
-                                                          device, dtype) for i in range(num_layers)])
-        self.num_layers = num_layers
-        self.norm = norm
-
-    def forward(self, src: Tensor, mask: Optional[Tensor] = None,
-                src_key_padding_mask: Optional[Tensor] = None) -> Tensor:
-        r"""Pass the input through the encoder layers in turn.
-
-        Args:
-            src: the sequence to the encoder (required).
-            mask: the mask for the src sequence (optional).
-            src_key_padding_mask: the mask for the src keys per batch (optional).
-
-        Shape:
-            see the docs in Transformer class.
-        """
-        output = src
-
-        for mod in self.layers:
-            output = mod(output, mask, src_key_padding_mask)
-
-        if self.norm is not None:
-            output = self.norm(output)
-
-        return output
-
-
-class TransformerDecoder(Module):
-    r"""TransformerDecoder is a stack of N decoder layers
-
-    Args:
-        decoder_layer: an instance of the TransformerDecoderLayer() class (required).
-        num_layers: the number of sub-decoder-layers in the decoder (required).
-        norm: the layer normalization component (optional).
-    """
-    __constants__ = ['norm']
-
-    #def __init__(self, decoder_layer, num_layers, norm=None):
-    def __init__(self, num_layers, d_model, nhead, dim_feedforward=2048, dropout=0.1, activation="relu",
-                 layer_norm_eps=1e-5, batch_first=False, norm_first=False, norm=None,
-                 device=None, dtype=None):
-        super(TransformerDecoder, self).__init__()
-        # self.layers = _get_clones(decoder_layer, num_layers)
-        self.layers = ModuleList([TransformerDecoderLayer(d_model, nhead, dim_feedforward, dropout,
-                                                          activation, layer_norm_eps, batch_first, norm_first,
-                                                          device, dtype) for i in range(num_layers)])
-        self.num_layers = num_layers
-        self.norm = norm
-
-    def forward(self, tgt: Tensor, memory: Tensor, tgt_mask: Optional[Tensor] = None,
-                memory_mask: Optional[Tensor] = None, tgt_key_padding_mask: Optional[Tensor] = None,
-                memory_key_padding_mask: Optional[Tensor] = None) -> Tensor:
-        r"""Pass the inputs (and mask) through the decoder layer in turn.
-
-        Args:
-            tgt: the sequence to the decoder (required).
-            memory: the sequence from the last layer of the encoder (required).
-            tgt_mask: the mask for the tgt sequence (optional).
-            memory_mask: the mask for the memory sequence (optional).
-            tgt_key_padding_mask: the mask for the tgt keys per batch (optional).
-            memory_key_padding_mask: the mask for the memory keys per batch (optional).
-
-        Shape:
-            see the docs in Transformer class.
-        """
-        output = tgt
-
-        for mod in self.layers:
-            output = mod(output, memory, tgt_mask, memory_mask,
-                         tgt_key_padding_mask, memory_key_padding_mask)
-
-        if self.norm is not None:
-            output = self.norm(output)
-
-        return output
-
-
-class TransformerEncoderLayer(Module):
-    r"""
-    Args:
-        d_model: the number of expected features in the input (required).
-        nhead: the number of heads in the multiheadattention models (required).
-        dim_feedforward: the dimension of the feedforward network model (default=2048).
-        dropout: the dropout value (default=0.1).
-        activation: the activation function of intermediate layer, relu or gelu (default=relu).
-        layer_norm_eps: the eps value in layer normalization components (default=1e-5).
-        batch_first: If ``True``, then the input and output tensors are provided
-            as (batch, seq, feature). Default: ``False``.
-        norm_first: if ``True``, layer norm is done prior to attention and feedforward
-            operations, respectivaly. Otherwise it's done after. Default: ``False`` (after).
-    """
-    __constants__ = ['batch_first', 'norm_first']
-
-    def __init__(self, d_model, nhead, dim_feedforward=2048, dropout=0.1, activation="relu",
-                 layer_norm_eps=1e-5, batch_first=False, norm_first=False,
-                 device=None, dtype=None) -> None:
-        #factory_kwargs = {'device': device, 'dtype': dtype}
-        super(TransformerEncoderLayer, self).__init__()
-        self.self_attn = MultiheadAttention(
-            d_model, nhead, dropout=dropout, batch_first=batch_first)
-        # Implementation of Feedforward model
-        self.linear1 = Linear(d_model, dim_feedforward)
-        self.dropout = Dropout(dropout)
-        self.linear2 = Linear(dim_feedforward, d_model)
-
-        self.norm_first = norm_first
-        self.norm1 = LayerNorm(d_model, eps=layer_norm_eps)
-        self.norm2 = LayerNorm(d_model, eps=layer_norm_eps)
-        self.dropout1 = Dropout(dropout)
-        self.dropout2 = Dropout(dropout)
-
-        self.activation = _get_activation_fn(activation)
-    ''' 
-    def __setstate__(self, state):
-        if 'activation' not in state:
-            state['activation'] = F.relu
-        super(TransformerEncoderLayer, self).__setstate__(state)
-    '''
-
-    def forward(self, src: Tensor, src_mask: Optional[Tensor] = None,
-                src_key_padding_mask: Optional[Tensor] = None) -> Tensor:
-        r"""Pass the input through the encoder layer.
-
-        Args:
-            src: the sequence to the encoder layer (required).
-            src_mask: the mask for the src sequence (optional).
-            src_key_padding_mask: the mask for the src keys per batch (optional).
-
-        Shape:
-            see the docs in Transformer class.
-        """
-        if self.norm_first:
-            src = self.norm1(src)
-            src2 = self.self_attn(
-                src, src, src, src_key_padding_mask, True, src_mask)[0]
-            src = src + self.dropout1(src2)
-            src = self.norm2(src)
-            src2 = self.linear2(self.dropout(
-                self.activation(self.linear1(src))))
-            src = src + self.dropout2(src2)
-            return src
-
-        # norm last
-        src2 = self.self_attn(
-            src, src, src, src_key_padding_mask, True, src_mask)[0]
-        src = src + self.dropout1(src2)
-        src = self.norm1(src)
-        src2 = self.linear2(self.dropout(self.activation(self.linear1(src))))
-        src = src + self.dropout2(src2)
-        src = self.norm2(src)
-        return src
-
-
-class TransformerDecoderLayer(Module):
-    r"""
-    Args:
-        d_model: the number of expected features in the input (required).
-        nhead: the number of heads in the multiheadattention models (required).
-        dim_feedforward: the dimension of the feedforward network model (default=2048).
-        dropout: the dropout value (default=0.1).
-        activation: the activation function of intermediate layer, relu or gelu (default=relu).
-        layer_norm_eps: the eps value in layer normalization components (default=1e-5).
-        batch_first: If ``True``, then the input and output tensors are provided
-            as (batch, seq, feature). Default: ``False``.
-        norm_first: if ``True``, layer norm is done prior to self attention, multihead
-            attention and feedforward operations, respectivaly. Otherwise it's done after.
-            Default: ``False`` (after).
-    """
-    __constants__ = ['batch_first', 'norm_first']
-
-    def __init__(self, d_model, nhead, dim_feedforward=2048, dropout=0.1, activation="relu",
-                 layer_norm_eps=1e-5, batch_first=False, norm_first=False,
-                 device=None, dtype=None) -> None:
-        #factory_kwargs = {'device': device, 'dtype': dtype}
-        super(TransformerDecoderLayer, self).__init__()
-        self.self_attn = MultiheadAttention(
-            d_model, nhead, dropout=dropout, batch_first=batch_first)
-        self.multihead_attn = MultiheadAttention(
-            d_model, nhead, dropout=dropout, batch_first=batch_first)
-        # Implementation of Feedforward model
-        self.linear1 = Linear(d_model, dim_feedforward)
-        self.dropout = Dropout(dropout)
-        self.linear2 = Linear(dim_feedforward, d_model)
-
-        self.norm_first = norm_first
-        self.norm1 = LayerNorm(d_model, eps=layer_norm_eps)
-        self.norm2 = LayerNorm(d_model, eps=layer_norm_eps)
-        self.norm3 = LayerNorm(d_model, eps=layer_norm_eps)
-        self.dropout1 = Dropout(dropout)
-        self.dropout2 = Dropout(dropout)
-        self.dropout3 = Dropout(dropout)
-
-        self.activation = _get_activation_fn(activation)
-    '''
-    def __setstate__(self, state):
-        if 'activation' not in state:
-            state['activation'] = F.relu
-        super(TransformerDecoderLayer, self).__setstate__(state)
-    '''
-
-    def forward(self, tgt: Tensor, memory: Tensor, tgt_mask: Optional[Tensor] = None,
-                memory_mask: Optional[Tensor] = None,
-                tgt_key_padding_mask: Optional[Tensor] = None,
-                memory_key_padding_mask: Optional[Tensor] = None) -> Tensor:
-        r"""Pass the inputs (and mask) through the decoder layer.
-
-        Args:
-            tgt: the sequence to the decoder layer (required).
-            memory: the sequence from the last layer of the encoder (required).
-            tgt_mask: the mask for the tgt sequence (optional).
-            memory_mask: the mask for the memory sequence (optional).
-            tgt_key_padding_mask: the mask for the tgt keys per batch (optional).
-            memory_key_padding_mask: the mask for the memory keys per batch (optional).
-
-        Shape:
-            see the docs in Transformer class.
-        """
-        if self.norm_first:
-            tgt = self.norm1(tgt)
-            tgt2 = self.self_attn(
-                tgt, tgt, tgt, tgt_key_padding_mask, True, tgt_mask)[0]
-            tgt = tgt + self.dropout1(tgt2)
-            tgt = self.norm2(tgt)
-            tgt2 = self.multihead_attn(
-                tgt, memory, memory, memory_key_padding_mask, True, memory_mask)[0]
-            tgt = tgt + self.dropout2(tgt2)
-            tgt = self.norm3(tgt)
-            tgt2 = self.linear2(self.dropout(
-                self.activation(self.linear1(tgt))))
-            tgt = tgt + self.dropout3(tgt2)
-            return tgt
-
-        # norm last
-        tgt2 = self.self_attn(
-            tgt, tgt, tgt, tgt_key_padding_mask, True, tgt_mask)[0]
-        tgt = tgt + self.dropout1(tgt2)
-        tgt = self.norm1(tgt)
-        tgt2 = self.multihead_attn(
-            tgt, memory, memory, memory_key_padding_mask, True, memory_mask)[0]
-        tgt = tgt + self.dropout2(tgt2)
-        tgt = self.norm2(tgt)
-        tgt2 = self.linear2(self.dropout(self.activation(self.linear1(tgt))))
-        tgt = tgt + self.dropout3(tgt2)
-        tgt = self.norm3(tgt)
-        return tgt
-
-# need deepcopy
-
-
-def _get_clones(module, N):
-    return ModuleList([copy.deepcopy(module) for i in range(N)])
-
-
-def _get_activation_fn(activation):
-    if activation == "relu":
-        return F.relu
-    elif activation == "gelu":
-        return F.gelu
-
-    raise RuntimeError(
-        "activation should be relu/gelu, not {}".format(activation))
