--- Transformer/requirements.txt
@@ -1,3 +1,0 @@
-numpy
-six
-shutil
