+++ imdb/README.md
@@ -1,0 +1,7 @@
+# IMDB 数据
+
+处理成 OFRecord 格式的 [IMDB](https://ai.stanford.edu/~amaas/data/sentiment/) 数据集 
+
+- train-part-0：训练数据集
+- test-part-0：测试数据集
+- word_index.json：词表
